Deepcopy values up front, not on each access
Deepcopying things repeatedly is a major performance issue.
Just doing it once is a huge speed and memory boost in case of
big blueprints.

It seems that the values are never mutated, so this optimization
is correct; this was checked by the unittests, and by inspecting
every use site of .initial_value and .value