CFY-7806 Batch dependencies between element types (#129)
When elements of type A must be parsed before elements of type B
(eg. relationships must be parsed before relationship properties),
we used to force that by adding a graph edge between every A and
every B.

However, that means we create a total of len(A)*len(B) edges, which
for 10k relationships means 10k*10k = 1e10 edges.

Instead, we create an artificial _BatchDependency() object, and
create edges between it and all A (for 10k edges), and between all B
and that object (another 10k edges, for a total of 20k).

This allows creating graphs of blueprints with a huge number of
relationships (on the order of 10k) without using up all available
memory.